Move gists/gist5639599 to printerutil.
